Test sandbox lifecycle callback
Use addSdkSandboxLifecycleCallback() to register for sandbox death event
notification. Additionally, update the SDK and client app with all the updated API changes.

Test: Run ./gradlew sdk-app:installDebug and ./gradlew
client-app:installDebug. In the app, load the sdk, and kill the sandbox
using `adb shell kill -9 <sandbox-pid>`. A toast appears on the app
notifying that the sandbox has died

Change-Id: I8839f536b84a7abb1fadca86a2baf5faa76189d0