Add aidl library developed by the sdk
The client-app depends on this library and uses it to
understand the binder object it receives from the Sdk as a response to
loadSdk.

This change shows a sample hello message being passed from the app to
the sdk in the privacy sandbox using the binder object sent as a part of
SandboxedSdk.

Test: Run ./gradlew sdk-app:installDebug and ./gradlew
client-app: installDebug
Change-Id: If4c46119e83c8752fe68b7c8812e53ccc0c9c5c9