Modified sample app to support frequency capping.
Modified ad selection wrapper logic to immeiately call
updateAdCounterHistogram with an impression event after a successful
adselection.

Created a custom audience that bids higher than all other CAs and is filtered after it is associated with
2 impression events or 1 click event. After 10 seconds, both caps are
  reset.

Added a manual button to update the histogram for a click event for a
given adselectionId.

Modified both override js and mock server js with new bidding logic.

Bug: 265204823
Change-Id: Ib63fb08dd0a6f9f9a346d5ae0293c21ccce4af35