Change Kotlin sample to use latests Privacy Sandbox libraries.
The sample is now...:
- using Privacy Sandbox tool libraries for code generation and API definition.
- using the UI library for remote rendering.
- using SDK Runtime library for backwards compatibility (not fully
  supported by Android Studio at the time, but moving forward the
  library will be the recommended way of interacting with the sandbox.)

The sample now also launches activities with the UI library and reads
resources from the Privacy Sandbox.

One drawback is that I had to remove the mediation examples from the
sample because Android Studio support for mediation is still under
development when using the generated sources from the Privacy Sandbox
tool annotations.

Change-Id: Ia2f914496d51b14c4066423cdecd666896c54d64