Add a wrapper SDK to sample app
The wrapper SDK represents a real SDK and calls the sandboxed SDK.

It represents a transition period where the client calls the statically linked SDK library as usual, while the SDK library starts to increasingly divert some requests to the runtime enabled SDK if available. This allows the SDK to incrementally move into the runtime.

Change-Id: I0f6895dcc6bf34823d2f6330a8d28ebe4d7ebd16